👑No Roles

+++ commands/admin/deleteRole.ts
@@ -15,6 +15,8 @@ const DeleteRoleCommand: Command = {
     async run(client, message, args) {
         const role = message.mentions.roles.first() || message.guild?.roles.cache.get(args[0])
 
+        if (!role) return message.channel.send("**Please specify a role to delete**")
+
         // @ts-ignore
         message.guild?.roles.cache.get(role?.id)?.delete()
